requirements.txt
django==1.9.9
bootstrap-admin==0.3.7.1
django-jsonfield==1.0.1
django-bootstrap3==7.1.0
django-phonenumber-field==1.1.0
psycopg2==2.6.2
django-mailgun==0.9.1
python-social-auth==0.2.21
django-filter==1.0.1
django-pagination-bootstrap==1.1
django-celery==3.1.17
celery==3.1.24
twilio==5.6.0
redis==2.10.5
beautifulsoup4==4.4.1
django-dbbackup==3.1.3
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
markdown==2.6.8
